use plugin
plugin is the set of custome rule that used in project.
plugin and configuration files  are used to set rules in project.
inorder to share the rules we use plugin and configuraion files to share and make other to use it .

we use third-party plugin to use in our project.we cant get it from npm and install it in project
then do changes in configration file .set "plugin":["plugin_name"];
and specified in that rule in rules object which rule we want to use .
here promiss plugin is used i.e eslint-plugin-promise .we can get it on npm website